move common log function out of processing.
- add typing for Loggers
-
