+++ backend/wmg/data/validation/validation.py
@@ -10,7 +10,7 @@ from backend.common.utils.math_utils import GB
 from backend.wmg.data.snapshot import CELL_COUNTS_CUBE_NAME, EXPRESSION_SUMMARY_CUBE_NAME
 from backend.wmg.data.validation import fixtures
 
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
 
 class Validation:
