+++ scripts/cxg_admin_scripts/reprocess_datafile.py
@@ -12,7 +12,7 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "..."))  # no
 sys.path.insert(0, pkg_root)  # noqa
 
 logging.basicConfig()
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
 
 def get_aws_account_id() -> str:
