+++ scripts/cxg_admin_scripts/updates.py
@@ -10,7 +10,7 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "..."))  # no
 sys.path.insert(0, pkg_root)  # noqa
 
 logging.basicConfig()
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
 auth0_apis = {
     "staging": "https://czi-cellxgene-dev.us.auth0.com",
