+++ tests/unit/backend/common/test_logger.py
@@ -1,6 +1,6 @@
 import pytest
 
-from backend.layers.processing.logger import LogSuppressed
+from backend.common.logging_config import LogSuppressed
 
 
 def test_log_suppressed__caught(caplog):
